Add MessageDTO and GameIncreasePopularityHandler
A new `MessageDTO` class was introduced to encapsulate user
and game-related data. This class includes `UserId` and
`GamesId` properties.

The `GameIncreasePopularityHandler` class was added to
handle messages of type `MessageDTO`. It uses the
`IGameService` to increase the popularity of specified
games and logs the processing details.

Updated `DependencyInjectionConfiguration.cs` to include
the `FCG.Games.Application.DTOs` namespace.